Added sharing

+++ server/routes/metric.js
@@ -1,13 +1,19 @@
 const express = require('express');
 
 const { addMetric, deleteMetric, editMetric, getAllReadingData } = require('../functions/metrics');
+const { addShare, deleteShare, editShare } = require('../functions/shares');
+const { canEditMetric, canViewMetric } = require('../functions/auth');
 const reading = require('./reading');
 
 const router = express.Router();
 
 router.get('/:metricId', async (req, res) => {
   try {
-    res.json(await getAllReadingData(req.params.metricId));
+    if (!await canViewMetric(req.trackId, req.params.metricId)) {
+      res.status(401).send('Invalid credentials');
+    } else {
+      res.json(await getAllReadingData(req.params.metricId));
+    }
   } catch (e) {
     res.status(400).send(e.message);
   }
@@ -25,9 +31,13 @@ router.post('/', async (req, res) => {
 
 router.patch('/:metricId', async (req, res) => {
   try {
-    const metric = await editMetric(req.params.metricId, req.body);
-    req.socketio.emit('metric_update', metric);
-    res.json(metric);
+    if (!await canEditMetric(req.trackId, req.params.metricId)) {
+      res.status(401).send('Invalid credentials');
+    } else {
+      const metric = await editMetric(req.params.metricId, req.body);
+      req.socketio.emit('metric_update', metric);
+      res.json(metric);
+    }
   } catch (e) {
     res.status(400).send(e.message);
   }
@@ -35,12 +45,56 @@ router.patch('/:metricId', async (req, res) => {
 
 router.delete('/:metricId', async (req, res) => {
   try {
-    const metric = await deleteMetric(req.params.metricId);
-    req.socketio.emit('metric_delete', {
-      track_id: req.trackId,
-      metric_id: req.params.metricId,
-    });
-    res.send(metric);
+    if (!await canEditMetric(req.trackId, req.params.metricId)) {
+      res.status(401).send('Invalid credentials');
+    } else {
+      const metric = await deleteMetric(req.params.metricId);
+      req.socketio.emit('metric_delete', {
+        track_id: req.trackId,
+        metric_id: req.params.metricId,
+      });
+      res.send(metric);
+    }
+  } catch (e) {
+    res.status(400).send(e.message);
+  }
+});
+
+router.post('/:metricId/shares', async (req, res) => {
+  try {
+    const share = await addShare(req.trackId, req.params.metricId, req.body)
+    req.socketio.emit('share_new', share);
+    res.json(share);
+  } catch (e) {
+    res.status(400).send(e.message);
+  }
+});
+
+router.patch('/:metricId/shares', async (req, res) => {
+  try {
+    const share = await editShare(req.trackId, req.params.metricId, req.body);
+    req.socketio.emit('share_update', share);
+    res.json(share);
+  } catch (e) {
+    res.status(400).send(e.message);
+  }
+});
+
+router.put('/:metricId/shares', async (req, res) => {
+  try {
+    const share = await editShare(req.trackId, req.params.metricId, req.body);
+    req.socketio.emit('share_accept', share);
+    res.json(share);
+  } catch (e) {
+    res.status(400).send(e.message);
+  }
+});
+
+router.delete('/:metricId/shares', async (req, res) => {
+  try {
+    const share = await deleteShare(req.trackId, req.params.metricId, req.body);
+    req.socketio.emit('share_delete', share);
+    res.send("Share successfully deleted");
   } catch (e) {
     res.status(400).send(e.message);
   }
